Move AWS image upload code to different action
